wordle: Go back to enum-based game states
Replace type-state driven game states with enum-backed states with
distinct structs for `PlayingGame` and `GameResult`.

Previously, the game was implemented in this style but it was
unergonomic to deal with because of some API decisions made. The
type-state pattern made for a slightly better experience, but also made
refactoring and plans to be able to serialize/deserialize games
difficult. Returning to an enum-based implementation with some of the
niceties of the newer typestate API achieves a good balance of both.