swordle: Optimize word constructor checks
Update `Word::new()'s` checks when ensuring the word is valid to only
binary search over the `GUESSES` constant instead of `SOLUTIONS` then
`GUESSES`. While this can be slightly slower when creating/validating a
solution word, in most cases it is actually faster since it only needs
to search over a single list of words.